Backend/requirements.txt
fastapi
sqlalchemy
python-dotenv
pydantic
pymupdf
google-generativeai
pickle-mixin
numpy
joblib
passlib
scikit-learn==1.3.2
mysql-connector-python
uvicorn[standard]
pandas
mlflow
dvc
prefect